refactor: Separate files by responsibility
